requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comtypes==1.2.0
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.12
multidict==6.0.4
PyAudio==0.2.14
pydantic==2.3.0
pydantic_core==2.6.3
pypiwin32==223
python-dotenv==1.0.0
pyttsx3==2.90
pywin32==306
requests==2.31.0
SpeechRecognition==3.10.0
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
